File distribution is done. Need to work on portfolio and main drawer props
