+++ src/javascript/app_2/components/elements/drawer/drawer_item.jsx
@@ -5,14 +5,14 @@ import { BinaryLink } from '../../../routes';
 
 class DrawerItem extends PureComponent {
 
-    onClick = () => {
-        this.console.log('hello world');
+    drawerItemClicked = () => {
+        console.log('Drawer Item clicked');
+        this.props.hideMainDrawer();
     }
 
     render() {
-        console.log(this.props);
         return (
-            <div className='drawer-item' onClick={()=>this.onClick()}>
+            <div className='drawer-item' onClick={()=>this.drawerItemClicked()}>
                 <BinaryLink to={this.props.link_to}>
                     <span className={this.props.icon || undefined}>{this.props.text}</span>
                 </BinaryLink>
@@ -30,11 +30,12 @@ DrawerItem.propTypes = {
     text: PropTypes.string,
 };
 
-const drawer_component = connect(
-    ({ ui }) => ({
-        is_portfolio_drawer_on: ui.is_portfolio_drawer_on,
-        togglePortfolioDrawer : ui.togglePortfolioDrawer,
+const drawer_item_component = connect(
+    ({ ui: {
+        is_main_drawer_on, toggleMainDrawer, hideMainDrawer
+    }}) => ({
+        is_main_drawer_on, toggleMainDrawer, hideMainDrawer
     })
 )(DrawerItem);
 
-export { drawer_component as DrawerItem };
+export { drawer_item_component as DrawerItem };
