+++ src/javascript/app_2/pages/trading/trade_app.jsx
@@ -10,7 +10,7 @@ import StartDate       from './components/start_date.jsx';
 import Symbol          from './components/symbol.jsx';
 import Test            from './components/test.jsx';
 import MobileWidget    from './components/elements/mobile_widget.jsx';
-import PortfolioDrawer from '../../components/elements/portfolio_drawer.jsx';
+import { PortfolioDrawer } from '../../components/elements/drawer';
 import { connect }     from '../../store/connect';
 
 const form_components = [
